Integrate RabbitMQ with connection management improvements
This commit introduces a structured approach to RabbitMQ integration by replacing direct connection handling with a connection factory and accessor. New classes (`RabbitMqConnectionAccessor`, `RabbitMqConnectionFactory`, and `RabbitMqConnectionInitializer`) have been added to manage RabbitMQ connections. The `SubscriptionService` has been updated to utilize the new factory, and the `EventBusProducer` now relies on the connection accessor. Logging has been enhanced with Serilog, and RabbitMQ configuration has been centralized in `Program.cs`.
